requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bs4==0.0.2
fastapi==0.115.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
imbalanced-learn==0.13.0
ipython==8.32.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
matplotlib==3.10.0
nltk==3.9.1
numpy==1.26.4
openai==1.59.9
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.10.5
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.28.1
sentence-transformers==3.3.1
spacy==3.8.4
SQLAlchemy==2.0.37
torch==2.5.1
transformers==4.48.0
Unidecode==1.3.8
Werkzeug==3.1.3
WTForms==3.2.1
xgboost==2.1.4
